pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "thesis-llm-benchmark"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "openai",
  "python-dotenv",
]
